yardman/requirements.txt
apturl==0.5.2
asgiref==3.2.5
autopilot==1.6.0
bcrypt==3.1.7
billiard==3.6.3.0
blinker==1.4
Brlapi==0.7.0
celery==5.0.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
datapunt-keycloak-oidc==0.5
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.4
django-actions-logger==0.3.1
django-adminlte2==0.4.1
django-ckeditor==6.0.0
django-cors-headers==3.2.1
django-environ==0.4.5
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-richtextfield==1.5.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
duplicity==0.8.12.0
ecdsa==0.14.1
entrypoints==0.3
environ==1.0
et-xmlfile==1.0.1
evdev==1.3.0
extras==1.0.0
fasteners==0.14.1
fixtures==3.0.0
future==0.18.2
GDAL==3.0.4
gunicorn==20.0.4
httplib2==0.14.0
idna==2.9
jdcal==1.4.1
josepy==1.3.0
junitxml==0.6
keyring==18.0.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linecache2==1.0.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
meld==3.20.2
monotonic==1.5
mozilla-django-oidc==1.2.3
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.5
paramiko==2.6.0
pbr==5.4.5
pdfkit==0.6.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.4
pyarmor==6.7.2
pyarmor-webui==1.2.9
pyasn1==0.4.8
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
pyee==8.1.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyppeteer==0.2.5
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-jose==3.2.0
python-keycloak==0.19.0
python-mimeparse==1.6.0
python-subunit==1.4.0
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.34
requests==2.23.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
systemd-python==234
testscenarios==0.5.0
testtools==2.4.0
traceback2==1.4.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
xkit==0.0.0
xlrd==2.0.1
